feat: Added tweening to score text

+++ Breakin/src/Tween/Tween.h
@@ -9,16 +9,16 @@ namespace bin
 {
     struct Tween
     {
-        double delay = 0.0;
-        double from = 0.0;
-        double to = 1.0;
-        double duration = 1.0;
+        float delay = 0.0;
+        float from = 0.0;
+        float to = 1.0;
+        float duration = 1.0;
         bool igunoreTimeScale = false;
         bool invokeWhenDestroyed = false;
         EaseType easeType = EaseType::Linear;
 
         std::function<void()> onStart = {};
-        std::function<void(double)> onUpdate = {};
+        std::function<void(float)> onUpdate = {};
         std::function<void()> onEnd = {};
     };
 }  // namespace jul
